with out application.properties

+++ src/main/java/pl/com/britenet/lambda/repository/UserRepository.java
@@ -3,5 +3,5 @@ package pl.com.britenet.lambda.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.com.britenet.lambda.entity.User;
 
-public interface UserRepository extends CrudRepository<User, Integer> {
+public interface UserRepository extends CrudRepository<User, Long> {
 }
